Renaming to prepare for public release v0.1.0. Tested packaging and upload the TestPyPI success.
