+++ pyproject.toml
@@ -1,0 +1,22 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "chatarena"
+version = "0.1.0"
+authors = [
+    { name = "Yuxiang Wu", email = "yuxiang.cs@gmail.com" },
+]
+description = "Multi-Agent Language Game Environments for LLMs"
+readme = "README.md"
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/chatarena/chatarena"
+"Bug Tracker" = "https://github.com/chatarena/chatarena/issues"
